requirements.txt
absl-py==1.2.0
alembic==1.8.1
aniso8601==9.0.1
astunparse==1.6.3
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.5
cycler==0.11.0
databricks-cli==0.17.3
docker==6.0.0
entrypoints==0.4
Flask==2.2.2
Flask-RESTful==0.3.9
flatbuffers==2.0.7
fonttools==4.37.3
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3
grpcio==1.48.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
mlflow==1.29.0
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.3
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
querystring-parser==1.2.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.41
sqlparse==0.4.2
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
typing_extensions==4.3.0
urllib3==1.26.12
websocket-client==1.4.1
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
